git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-sasl-branch@541123 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/directory/server/core/partition/impl/btree/BTreeSearchResultEnumeration.java
@@ -104,11 +104,11 @@ public class BTreeSearchResultEnumeration implements SearchResultEnumeration
     {
         IndexRecord rec = ( IndexRecord ) underlying.next();
         Attributes entry;
-        String name = partition.getEntryUpdn( rec.getEntryId() );
+        String name = partition.getEntryUpdn( (Long)rec.getEntryId() );
 
         if ( null == rec.getAttributes() )
         {
-            rec.setAttributes( partition.lookup( rec.getEntryId() ) );
+            rec.setAttributes( partition.lookup( (Long)rec.getEntryId() ) );
         }
 
         if ( attrIds == null )
@@ -234,7 +234,7 @@ public class BTreeSearchResultEnumeration implements SearchResultEnumeration
             }
         }
 
-        BTreeSearchResult result = new BTreeSearchResult( rec.getEntryId(), name, null, entry );
+        BTreeSearchResult result = new BTreeSearchResult( (Long)rec.getEntryId(), name, null, entry );
         result.setRelative( false );
         return result;
     }
